Moved and updated all electrum-based methods from wallet/index and moved them to wallet/electrum.ts.
